# api/openapi.yaml
components:
  schemas:
    ActiveMethodView:
      description: 'One active authentication method instance. `id` addresses it for DELETE .../methods/{id} - method name
        alone isn''t unique when a method allows multiple instances (e.g. several active `device` entries, one per physical
        device). `label` is a user-chosen display name, set only for multi-instance methods; `null` for singleton ones (email/sms/password),
        which the client labels from `method` itself. `enrolledUnderAcr`/`maxAcr`/`effectiveAcr`/`factorTypes` surface the
        ADR-5 three-way cap (docs/12-entscheidungen.md): `effectiveAcr` is `min(enrolledUnderAcr, maxAcr)`, the level this
        method can actually contribute right now, which can be lower than the tool''s own declared `maxAcr` if it was enrolled
        while the session had proven less.'
      properties:
        effectiveAcr:
          description: min(enrolledUnderAcr, maxAcr) - what this method actually contributes today.
          example: loa1
          type: string
        enrolledUnderAcr:
          description: The level the session had already proven at the moment this method was enrolled (ADR-5) - caps effectiveAcr
            below maxAcr if lower.
          example: loa1
          type: string
        factorTypes:
          example:
          - POSSESSION
          items:
            enum:
            - KNOWLEDGE
            - POSSESSION
            - INHERENCE
            type: string
          type: array
        id:
          example: 7f3e2b1a-0c9d-4e8f-8a1b-2c3d4e5f6a7b
          type: string
        label:
          example: Laptop
          type: string
        maxAcr:
          description: This tool's own declared ceiling - never account/session-specific.
          example: loa2
          type: string
        method:
          example: sms
          type: string
      required:
      - id
      - method
      type: object
...
    AuthData:
      properties:
        accountId:
          example: 42
          format: int64
          type: integer
        acr:
          example: loa2
          type: string
        amr:
          additionalProperties:
            type: string
          description: 'Method -> who proved it: "orchestrator" for a completed orchestrator tool, "kc" for evidence a native
            Keycloak authenticator already established (docs/05-api.md Abschnitt 3). Informational only - the orchestrator
            alone still resolves the combined acr above, regardless of source.'
          example:
            password: kc
            sms: orchestrator
          type: object
      type: object
...
    ChannelBlock:
      properties:
        activeMethods:
          description: All active authentication methods on the account, regardless of whether this session's currentAmr proved
            them. currentAmr is session evidence (what THIS channel actually proved); activeMethods is the account's full
            standing method list, unfiltered by device - a lost/stolen device's credential must be removable from any authenticated
            session, not only from that device itself.
          items:
            $ref: '#/components/schemas/ActiveMethodView'
          type: array
        channelSessionId:
          format: uuid
          type: string
        channelType:
          description: Which facade this channel was opened through - APP (DPoP) or KEYCLOAK (docs/02-domaenenmodell.md Abschnitt
            1). Fixed for the channel's whole lifetime.
          example: APP
          type: string
        currentAcr:
          example: loa2
          type: string
        currentAmr:
          example:
          - sms
          - password
          items:
            type: string
          type: array
        state:
          example: AUTHENTICATED
          type: string
      required:
      - channelSessionId
      - channelType
      - state
      type: object
...
    ChannelResponse:
      properties:
        authData:
          $ref: '#/components/schemas/AuthData'
          description: KEYCLOAK channels only (docs/05-api.md Abschnitt 3) - never present for APP.
        channel:
          $ref: '#/components/schemas/ChannelBlock'
        demo:
          $ref: '#/components/schemas/DemoInfo'
          description: Demo-only correlation IDs, never part of the production contract.
        next:
          $ref: '#/components/schemas/Next'
        stepData:
          additionalProperties:
            type: object
          description: 'Whatever the current step needs to render: missing fields, selection options, or a retry reason.'
          example:
            options:
            - ident-fsc
            - ident-eid
          type: object
      required:
      - channel
      type: object
...
    ConfirmQrLoginActivateRequest:
      properties:
        pairingCode:
          example: ABCD-1234
          type: string
      type: object
...
    DemoInfo:
      additionalProperties: true
      properties:
        accountId:
          example: 42
          format: int64
          type: integer
        journeys:
          description: The running journey chain for this channel, outermost first - see [JourneyDebugStep]. Empty once nothing
            is running.
          items:
            $ref: '#/components/schemas/JourneyDebugStep'
          type: array
        personId:
          example: 7
          format: int64
          type: integer
      type: object
...
    JourneyDebugStep:
      properties:
        intent:
          example: DELETE_ACCOUNT
          type: string
        journeyId:
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          type: string
        lifecycle:
          example: SUSPENDED
          type: string
        note:
          description: 'Demo-only: why this journey''s current step looks the way it does - either why its tool became the
            automatic choice, or why a selection among several is being shown at all. Null whenever the step already explains
            itself (e.g. a Prompt), never part of the production contract.'
          example: Gerät wiedererkannt - automatisch vorgeschlagen.
          type: string
        stateType:
          example: ConfirmPending
          type: string
      required:
      - intent
      - journeyId
      - lifecycle
      - stateType
      type: object
...
    KobilActivateRequest:
      properties:
        activationCode:
          type: string
        pin:
          type: string
        tenantId:
          type: string
        userId:
          type: string
      required:
      - activationCode
      - pin
      - tenantId
      - userId
      type: object
...
    Next:
      properties:
        context:
          example: authentication
          type: string
        step:
          example: auth
          type: string
        toolId:
          example: auth-sms
          type: string
        toolSessionId:
          format: uuid
          type: string
        type:
          example: tool
          type: string
      required:
      - step
      - type
      type: object
...
  securitySchemes:
    dpop:
      description: DPoP proof JWT (docs/09-dpop.md); required on every App-facade call.
      in: header
      name: DPoP
      type: apiKey
info:
  description: App-facing (orchestrator-first) API for the DPoP-bound registration and login demo. See docs/05-api.md for
    the full contract; the Keycloak-facing kc facade is out of scope for this build (docs/08-projektrahmen.md).
  title: DPoP-Demo Orchestrator API
  version: v1
openapi: 3.1.0
paths:
...
  /mock-kobil/activate:
    post:
      description: Binds this device to the user and creates the device identifier. The activation code is spent in the process.
      operationId: activate_21
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KobilActivateRequest'
        required: true
      responses:
        '200':
          content:
            '*/*':
              schema:
                additionalProperties:
                  type: string
                type: object
          description: OK
      summary: MC SDK ActivateEvent
      tags:
      - Mock KOBIL
...
  /orchestrator/api/v1/channels/{channelSessionId}/tools/auth-device:
    post:
      description: 'No request body: toolId already carries kind and method.'
      operationId: activate_20
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      - in: query
        name: bindingKeyRef
        required: true
        schema:
          type: string
      responses:
        '201':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  state: ANONYMOUS
                next:
                  step: auth
                  toolId: auth-device
                  toolSessionId: 9c858901-8a57-4791-81fe-4c455b099bc9
                  type: tool
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Created
      security:
      - dpop: [
          ]
      summary: Activate auth-device
      tags:
      - 'Tool: Gerät'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/auth-email:
    post:
      description: 'No request body: toolId already carries kind and method.'
      operationId: activate_18
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      - in: query
        name: bindingKeyRef
        required: true
        schema:
          type: string
      responses:
        '201':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  currentAcr: loa1
                  state: STEP_UP_IN_PROGRESS
                next:
                  step: auth
                  toolId: auth-email
                  toolSessionId: 9c858901-8a57-4791-81fe-4c455b099bc9
                  type: tool
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Created
      security:
      - dpop: [
          ]
      summary: Activate auth-email
      tags:
      - 'Tool: E-Mail'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/auth-email-lookup:
    post:
      description: 'No request body: toolId already carries kind and method.'
      operationId: activate_19
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      - in: query
        name: bindingKeyRef
        required: true
        schema:
          type: string
      responses:
        '201':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  state: REGISTERING
                next:
                  step: auth
                  toolId: auth-email-lookup
                  toolSessionId: 9c858901-8a57-4791-81fe-4c455b099bc9
                  type: tool
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Created
      security:
      - dpop: [
          ]
      summary: Activate auth-email-lookup
      tags:
      - 'Tool: E-Mail'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/auth-kobil:
    post:
      operationId: activate_17
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      - in: query
        name: bindingKeyRef
        required: true
        schema:
          type: string
      responses:
        '201':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  state: ANONYMOUS
                next:
                  step: unlock
                  toolId: auth-kobil
                  toolSessionId: 9c858901-8a57-4791-81fe-4c455b099bc9
                  type: tool
                stepData:
                  unlockOptions:
                  - biometric
                  - password
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Created
      security:
      - dpop: [
          ]
      summary: Activate auth-kobil
      tags:
      - 'Tool: KOBIL'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/auth-password:
    post:
      description: 'No request body: toolId already carries kind and method.'
      operationId: activate_15
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      - in: query
        name: bindingKeyRef
        required: true
        schema:
          type: string
      responses:
        '201':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  currentAcr: loa1
                  state: STEP_UP_IN_PROGRESS
                next:
                  step: auth
                  toolId: auth-password
                  toolSessionId: 9c858901-8a57-4791-81fe-4c455b099bc9
                  type: tool
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Created
      security:
      - dpop: [
          ]
      summary: Activate auth-password
      tags:
      - 'Tool: Passwort'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/auth-password-lookup:
    post:
      description: 'No request body: toolId already carries kind and method.'
      operationId: activate_16
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      - in: query
        name: bindingKeyRef
        required: true
        schema:
          type: string
      responses:
        '201':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  state: REGISTERING
                next:
                  step: auth
                  toolId: auth-password-lookup
                  toolSessionId: 9c858901-8a57-4791-81fe-4c455b099bc9
                  type: tool
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Created
      security:
      - dpop: [
          ]
      summary: Activate auth-password-lookup
      tags:
      - 'Tool: Passwort'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/auth-qr:
    post:
      description: 'No request body: toolId already carries kind and method.'
      operationId: activate_13
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      - in: query
        name: bindingKeyRef
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: OK
      security:
      - dpop: [
          ]
      summary: Activate auth-qr
      tags:
      - 'Tool: QR-Login'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/auth-qr-lookup:
    post:
      description: 'No request body: toolId already carries kind and method.'
      operationId: activate_14
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      - in: query
        name: bindingKeyRef
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: OK
      security:
      - dpop: [
          ]
      summary: Activate auth-qr-lookup
      tags:
      - 'Tool: QR-Login'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/auth-sms:
    post:
      description: 'No request body: toolId already carries kind and method.'
      operationId: activate_11
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      - in: query
        name: bindingKeyRef
        required: true
        schema:
          type: string
      responses:
        '201':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  currentAcr: loa1
                  state: STEP_UP_IN_PROGRESS
                next:
                  step: auth
                  toolId: auth-sms
                  toolSessionId: 9c858901-8a57-4791-81fe-4c455b099bc9
                  type: tool
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Created
      security:
      - dpop: [
          ]
      summary: Activate auth-sms
      tags:
      - 'Tool: SMS'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/auth-sms-lookup:
    post:
      description: 'No request body: toolId already carries kind and method.'
      operationId: activate_12
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      - in: query
        name: bindingKeyRef
        required: true
        schema:
          type: string
      responses:
        '201':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  state: REGISTERING
                next:
                  step: auth
                  toolId: auth-sms-lookup
                  toolSessionId: 9c858901-8a57-4791-81fe-4c455b099bc9
                  type: tool
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Created
      security:
      - dpop: [
          ]
      summary: Activate auth-sms-lookup
      tags:
      - 'Tool: SMS'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/confirm-email:
    post:
      description: 'No request body: toolId already carries kind and method.'
      operationId: activate_10
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      - in: query
        name: bindingKeyRef
        required: true
        schema:
          type: string
      responses:
        '201':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  state: REGISTERING
                next:
                  step: input
                  toolId: confirm-email
                  toolSessionId: 9c858901-8a57-4791-81fe-4c455b099bc9
                  type: tool
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Created
      security:
      - dpop: [
          ]
      summary: Activate confirm-email
      tags:
      - 'Tool: E-Mail'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/confirm-qr-login:
    post:
      description: 'Optional body: {pairingCode}, when already known (e.g. from a demo-link deep link) - skips the input step.'
      operationId: activate_9
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      - in: query
        name: bindingKeyRef
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmQrLoginActivateRequest'
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: OK
      security:
      - dpop: [
          ]
      summary: Activate confirm-qr-login
      tags:
      - 'Tool: QR-Login'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/enroll-device:
    post:
      description: 'No request body: toolId already carries kind and method.'
      operationId: activate_8
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      - in: query
        name: bindingKeyRef
        required: true
        schema:
          type: string
      responses:
        '201':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  currentAcr: loa1
                  currentAmr:
                  - password
                  state: AUTHENTICATED
                next:
                  step: enroll
                  toolId: enroll-device
                  toolSessionId: 9c858901-8a57-4791-81fe-4c455b099bc9
                  type: tool
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Created
      security:
      - dpop: [
          ]
      summary: Activate enroll-device
      tags:
      - 'Tool: Gerät'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/enroll-email:
    post:
      description: 'One shot: no request body, and the response already carries the completed outcome.'
      operationId: activate_7
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      - in: query
        name: bindingKeyRef
        required: true
        schema:
          type: string
      responses:
        '201':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  state: AUTHENTICATED
                next:
                  context: authentication
                  step: done
                  type: orchestrator
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Created
      security:
      - dpop: [
          ]
      summary: Activate enroll-email
      tags:
      - 'Tool: E-Mail-Login'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/enroll-kobil:
    post:
      description: Provisions the KOBIL user and returns everything the app's SDK needs for its activation call - including
        the PIN, which this backend mints and keeps.
      operationId: activate_6
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      - in: query
        name: bindingKeyRef
        required: true
        schema:
          type: string
      responses:
        '201':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  currentAcr: loa2
                  currentAmr:
                  - fsc
                  state: AUTHENTICATED
                next:
                  step: activate
                  toolId: enroll-kobil
                  toolSessionId: 9c858901-8a57-4791-81fe-4c455b099bc9
                  type: tool
                stepData:
                  activationCode: K7M2PQX9
                  kobilUserId: kob-1a2b3c4d5e6f
                  missingFields:
                  - activated
                  - biometricConsent
                  pin: '40318827'
                  tenantId: dpop-demo
                  unlockSecret: xE1r...
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Created
      security:
      - dpop: [
          ]
      summary: Activate enroll-kobil
      tags:
      - 'Tool: KOBIL'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/enroll-password:
    post:
      description: 'No request body: toolId already carries kind and method.'
      operationId: activate_5
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      - in: query
        name: bindingKeyRef
        required: true
        schema:
          type: string
      responses:
        '201':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  state: REGISTERING
                next:
                  step: enroll
                  toolId: enroll-password
                  toolSessionId: 9c858901-8a57-4791-81fe-4c455b099bc9
                  type: tool
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Created
      security:
      - dpop: [
          ]
      summary: Activate enroll-password
      tags:
      - 'Tool: Passwort'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/enroll-qr:
    post:
      description: 'No request body: toolId already carries kind and method.'
      operationId: activate_4
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      - in: query
        name: bindingKeyRef
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: OK
      security:
      - dpop: [
          ]
      summary: Activate enroll-qr
      tags:
      - 'Tool: QR-Login'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/enroll-sms:
    post:
      description: 'No request body: toolId already carries kind and method.'
      operationId: activate_3
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      - in: query
        name: bindingKeyRef
        required: true
        schema:
          type: string
      responses:
        '201':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  state: REGISTERING
                next:
                  step: enroll
                  toolId: enroll-sms
                  toolSessionId: 9c858901-8a57-4791-81fe-4c455b099bc9
                  type: tool
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Created
      security:
      - dpop: [
          ]
      summary: Activate enroll-sms
      tags:
      - 'Tool: SMS'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/ident-eid:
    post:
      description: 'No request body: toolId already carries kind and method.'
      operationId: activate_2
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      - in: query
        name: bindingKeyRef
        required: true
        schema:
          type: string
      responses:
        '201':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  state: REGISTERING
                next:
                  step: card
                  toolId: ident-eid
                  toolSessionId: 9c858901-8a57-4791-81fe-4c455b099bc9
                  type: tool
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Created
      security:
      - dpop: [
          ]
      summary: Activate ident-eid
      tags:
      - 'Tool: eID'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/ident-fsc:
    post:
      description: 'No request body: toolId already carries kind and method.'
      operationId: activate_1
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      - in: query
        name: bindingKeyRef
        required: true
        schema:
          type: string
      responses:
        '201':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  state: REGISTERING
                next:
                  step: input
                  toolId: ident-fsc
                  toolSessionId: 9c858901-8a57-4791-81fe-4c455b099bc9
                  type: tool
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Created
      security:
      - dpop: [
          ]
      summary: Activate ident-fsc
      tags:
      - 'Tool: Freischaltcode'
  /orchestrator/api/v1/channels/{channelSessionId}/tools/ident-kvnr:
    post:
      description: 'No request body: toolId already carries kind and method.'
      operationId: activate
      parameters:
      - in: path
        name: channelSessionId
        required: true
        schema:
          format: uuid
          type: string
      - in: query
        name: bindingKeyRef
        required: true
        schema:
          type: string
      responses:
        '201':
          content:
            application/json:
              example:
                channel:
                  channelSessionId: 3fa85f64-5717-4562-b3fc-2c963f66afa6
                  state: REGISTERING
                next:
                  step: input
                  toolId: ident-kvnr
                  toolSessionId: 9c858901-8a57-4791-81fe-4c455b099bc9
                  type: tool
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: Created
      security:
      - dpop: [
          ]
      summary: Activate ident-kvnr
      tags:
      - 'Tool: Versichertennummer'
...
  /orchestrator/api/v1/tools/{toolSessionId}/enroll-email:
    get:
      operationId: read_20
      parameters:
      - in: path
        name: toolSessionId
        required: true
        schema:
          format: uuid
          type: string
      - in: query
        name: bindingKeyRef
        required: true
        schema:
          type: string
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ChannelResponse'
          description: OK
      security:
      - dpop: [
          ]
      summary: Read the current enroll-email state
      tags:
      - 'Tool: E-Mail-Login'
...